+++ crates/abd-clam/src/chaoda/graph/mod.rs
@@ -4,6 +4,6 @@ mod _graph;
 mod criteria;
 mod vertex;
 
-pub use _graph::{Edge, Graph};
+pub use _graph::{Edge, EdgeSet, Graph, VertexSet};
 pub use criteria::MetaMLScorer;
 pub use vertex::{Ratios, Vertex};
