+++ crates/abd-clam/src/chaoda/mod.rs
@@ -1,13 +1,12 @@
 //! TODO: Add more details on each modules
 
 // mod _chaoda;
-/// Modules related to various aspects of machine learning and data analysis.
 pub mod automl_regressors;
-/// Different ways to score clustered graphs based on their properties.
+pub mod graph;
 pub mod graph_scorers;
-/// A trait to be implemented by machine learning regressors.
 pub mod metaml;
-/// Contains pre-trained models for scoring clusters in graphs.
 pub mod pretrained_models;
+
+pub use graph::{Ratios, Vertex};
 
 // pub use _chaoda::CHAODA;
